+++ packages/orq-rc/RELEASES.md
@@ -5639,3 +5639,13 @@ Based on:
 - [typescript v4.2.0-rc.3] packages/orq-rc
 ### Releases
 - [NPM v4.2.0-rc.3] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.3 - packages/orq-rc
+
+## 2026-01-01 05:35:15
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.680.11 (2.788.15) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.2.0-rc.4] packages/orq-rc
+### Releases
+- [NPM v4.2.0-rc.4] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.4 - packages/orq-rc